+++ lib/apollo-federation/has_directives.rb
@@ -2,6 +2,8 @@
 
 module ApolloFederation
   module HasDirectives
+    attr_reader :federation_directives
+
     def add_directive(name:, arguments: nil)
       @federation_directives ||= []
       @federation_directives << { name: name, arguments: arguments }